review: 后端整数 rank 按宽度对齐前端, Byte↔Int8 提升; 修 Float/Float64 精确性检查恒 false; IntLit raw 符号感知; 32位格式化可移植; init 失败释放 builder

+++ frontend/src/cwind_frontend/sa/smt.py
@@ -407,7 +407,7 @@ class BodyChecks:
             folded = int(folded)
         if isinstance(folded, int):
             f32 = struct.unpack("!f", struct.pack("!f", float(folded)))[0]
-            if float(f32) != float(folded):
+            if int(f32) != folded:
                 self._record_error(
                     f"value {folded} is not exactly representable in Float",
                     value.line,
@@ -435,7 +435,7 @@ class BodyChecks:
             folded = int(folded)
         if isinstance(folded, int):
             f64 = struct.unpack("!d", struct.pack("!d", float(folded)))[0]
-            if float(f64) != float(folded):
+            if int(f64) != folded:
                 self._record_error(
                     f"value {folded} is not exactly representable in Float64",
                     value.line,
